refactor(bootstrapTests): extract drivers from supports
