+++ packages/vats/test/bootstrapTests/supports.js
@@ -3,6 +3,7 @@
 import { Fail } from '@agoric/assert';
 import { buildSwingset } from '@agoric/cosmic-swingset/src/launch-chain.js';
 import { BridgeId, VBankAccount } from '@agoric/internal';
+import { unmarshalFromVstorage } from '@agoric/internal/src/lib-chainStorage.js';
 import {
   makeFakeStorageKit,
   slotToRemotable,
@@ -14,7 +15,6 @@ import { E } from '@endo/eventual-send';
 import { makeQueue } from '@endo/stream';
 import { promises as fs } from 'fs';
 import { resolve as importMetaResolve } from 'import-meta-resolve';
-import { unmarshalFromVstorage } from '@agoric/internal/src/lib-chainStorage.js';
 import { boardSlottingMarshaller } from '../../tools/board-utils.js';
 
 // to retain for ESlint, used by typedef
@@ -167,123 +167,6 @@ export const makeRunUtils = (controller, log = (..._) => {}) => {
     });
 
   return harden({ runThunk, EV });
-};
-
-/**
- * @param {ReturnType<typeof makeRunUtils>} runUtils
- * @param {import('@agoric/internal/src/storage-test-utils.js').FakeStorageKit} storage
- * @param {import('../../tools/board-utils.js').AgoricNamesRemotes} agoricNamesRemotes
- */
-export const makeWalletFactoryDriver = async (
-  runUtils,
-  storage,
-  agoricNamesRemotes,
-) => {
-  const { EV } = runUtils;
-
-  const walletFactoryStartResult = await EV.vat('bootstrap').consumeItem(
-    'walletFactoryStartResult',
-  );
-  const bankManager = await EV.vat('bootstrap').consumeItem('bankManager');
-  const namesByAddressAdmin = await EV.vat('bootstrap').consumeItem(
-    'namesByAddressAdmin',
-  );
-
-  const marshaller = boardSlottingMarshaller(slotToRemotable);
-
-  /**
-   * @param {string} walletAddress
-   * @param {import('@agoric/smart-wallet/src/smartWallet.js').SmartWallet} walletPresence
-   * @param {boolean} isNew
-   */
-  const makeWalletDriver = (walletAddress, walletPresence, isNew) => ({
-    isNew,
-
-    /**
-     * @param {import('@agoric/smart-wallet/src/offers.js').OfferSpec} offer
-     * @returns {Promise<void>}
-     */
-    executeOffer(offer) {
-      const offerCapData = marshaller.toCapData(
-        harden({
-          method: 'executeOffer',
-          offer,
-        }),
-      );
-      return EV(walletPresence).handleBridgeAction(offerCapData, true);
-    },
-    /**
-     * @param {import('@agoric/smart-wallet/src/offers.js').OfferSpec} offer
-     * @returns {Promise<void>}
-     */
-    sendOffer(offer) {
-      const offerCapData = marshaller.toCapData(
-        harden({
-          method: 'executeOffer',
-          offer,
-        }),
-      );
-
-      return EV.sendOnly(walletPresence).handleBridgeAction(offerCapData, true);
-    },
-    tryExitOffer(offerId) {
-      const capData = marshaller.toCapData(
-        harden({
-          method: 'tryExitOffer',
-          offerId,
-        }),
-      );
-      return EV(walletPresence).handleBridgeAction(capData, true);
-    },
-    /**
-     * @template {(brands: Record<string, Brand>, ...rest: any) => import('@agoric/smart-wallet/src/offers.js').OfferSpec} M offer maker function
-     * @param {M} makeOffer
-     * @param {Parameters<M>[1]} firstArg
-     * @param {Parameters<M>[2]} [secondArg]
-     * @returns {Promise<void>}
-     */
-    executeOfferMaker(makeOffer, firstArg, secondArg) {
-      const offer = makeOffer(agoricNamesRemotes.brand, firstArg, secondArg);
-      return this.executeOffer(offer);
-    },
-    /**
-     * @template {(brands: Record<string, Brand>, ...rest: any) => import('@agoric/smart-wallet/src/offers.js').OfferSpec} M offer maker function
-     * @param {M} makeOffer
-     * @param {Parameters<M>[1]} firstArg
-     * @param {Parameters<M>[2]} [secondArg]
-     * @returns {Promise<void>}
-     */
-    sendOfferMaker(makeOffer, firstArg, secondArg) {
-      const offer = makeOffer(agoricNamesRemotes.brand, firstArg, secondArg);
-      return this.sendOffer(offer);
-    },
-    /**
-     * @returns {import('@agoric/smart-wallet/src/smartWallet.js').UpdateRecord}
-     */
-    getLatestUpdateRecord() {
-      const fromCapData = (...args) =>
-        Reflect.apply(marshaller.fromCapData, marshaller, args);
-      return unmarshalFromVstorage(
-        storage.data,
-        `published.wallet.${walletAddress}`,
-        fromCapData,
-      );
-    },
-  });
-
-  return {
-    /**
-     * @param {string} walletAddress
-     */
-    async provideSmartWallet(walletAddress) {
-      const bank = await EV(bankManager).getBankForAddress(walletAddress);
-      return EV(walletFactoryStartResult.creatorFacet)
-        .provideSmartWallet(walletAddress, bank, namesByAddressAdmin)
-        .then(([walletPresence, isNew]) =>
-          makeWalletDriver(walletAddress, walletPresence, isNew),
-        );
-    },
-  };
 };
 
 export const getNodeTestVaultsConfig = async (
